# asm: reuses the slice on SetBranchTargetOnNextNodes